Set the window background based on system theme
