+++ src/renderer/styles/theme.ts
@@ -1,3 +1,4 @@
+import { lightTheme as light, darkTheme as dark } from "common/theme";
 import { Theme } from "@emotion/styled";
 
 // Re-export the theme
@@ -8,25 +9,6 @@ export enum ThemeMode {
   LIGHT = "light",
 }
 
-export const lightTheme: Theme = {
-  primary: "#FFF",
-  secondary: "#000",
-  foreground: "#333",
-  foreground2: "#111329",
-  foreground3: "#ddd",
-  background: "#FAFAFA",
-  background2: "#F9FAFB",
-  background3: "#d4d4d5",
-};
-
-// Color palette: https://paletton.com/#uid=13X0u0k8Zaw6JmT8khnfM8Ll-7w
-export const darkTheme: Theme = {
-  primary: "#000",
-  secondary: "#FFF",
-  foreground: "#cbcbce",
-  foreground2: "#67697E",
-  foreground3: "#4A4C60",
-  background: "#22222C",
-  background2: "#1C1D30",
-  background3: "#111329",
-};
+// Re-export the common theme as the @emotion/styled theme
+export const lightTheme = light as Theme;
+export const darkTheme = dark as Theme;
